backend/requirements.txt
Django==5.2.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
django-cors-headers==4.7.0
django-q2==1.9.0
channels==4.2.2
requests==2.32.5
psycopg2-binary==2.9.10
pillow==11.2.1
openpyxl==3.1.5
